Featured Images: De-deplicated featured images within content.
Some sites have the featured image in the content because their existing theme doesn't support featured images properly or in a satisfactory way. This leads to the image appearing twice in Newspack because the theme correctly shows the image that also then appears in the content.

This command detects that situation with each post and corrects it by removing the featured image from the content.