Link new Guest Authors to original CPTs after conversion.
This change adds a new meta field ('_post_migrated_from') to newly created guest authors. The value is the ID of the post that the guest author was converted from when using the CPT -> GA conversion script. This allows other tooling to attribute posts to the new guest authors based on the CPT they used to be assigned to.