Changes how changes made by update_blocks_ids are logged
Previous version of this method returned an array with all the changes made,
so that the caller command might log them. This didn't work, because these
changes can be huge. Now forwarding an optional param to this method, a
log file, so that the method can itself log all the changes to it, and
doesn't have to return a huge array any longer.