feat: needed two additional scripts to get EC's missing posts on their DB.
EnergeticCity provided us with an XML export of their posts. It contained posts between 2015 and 2019 which they would like on their site, since we weren't able to scrape them. I had to perform a certain portion of this migration locally. Another portion of the migration now requires to be executed on their live site.