feat: created a function to seek missing images in <img> tags.
This function will look at a site's wp_posts table for any posts with <img> tags in their post_content field. Out of the posts that it does find, it will load the post_content as HTML and cycle through any <img> tags. If it cannot find the indicated file in the src attribute, it will attempt to either look for it locally in the wp-content/uploads folder, or attempt to download it. If the image is downloaded, it will attach it to the post, and update the HTML content. The HTML content is also updated if the file is found locally as well.