feat: obtaining the corresponding post ID in the new DB table.
The XML export that was provided is referencing old database IDs for it's featured images. With this implementation I loaded a previous table backup of wp_posts and I use that to find the GUID column for the original attachment row. Once I have that, I am able to identify the new attachment row in the new wp_posts table.