feat: created an automated process to remove dupe data from meta table.
This function was created to help automate the reduction of a large wp_postmeta table which could adversely affect performance. It does this by looking at every Post ID in wp_postmeta and seeing if there's a difference between the total row count and the unique meta_keys. If there is a difference, it then creates a dynamic key by concatenating meta_key and meta_value, and deleting any rows where the dynamic key has more than 1 row existing in the table.