feat: updating query so that all categories and post_tags are shown.
Previously, if a category/post_tag existed, but wasn't linked to any post, this query would ignore those records. This is an issue if those taxonomies have an non-zero count, as they would not be presented to the user as needing to be updated.

With this commit, if the taxonomy does not have any linked posts, the `real_count` (real count of linked posts to a particular taxonomy) column will display a 0.