feat: refactored code to account for posts with multiple featured imgs.
I delete the featured images set for posts that have an image at the beginning of their post. This is "low hanging fruit" because even if we were to figure out which featured image is the "correct" one, it would still have to be deleted because of the image at the beginning of the post.