fix: new command which addresses duplicate articles which were imported.
The meta_values and post_ids in $target_data were found to be duplicate data. This command will look at the first and second post those IDs point to and copy over certain data points to the first post (post_date, post_date_gmt, and post_content). Then it will proceed to clean up the meta data by deleting the duplicate meta data point and ensuring that the first post is the only one that is pointed to using the `newspack_original_article_id` meta_key.